Make the foxglove a private dep
